Solver state
